internal: Remove unqualified_path completions module
